[cpu] combine memorymap.rs into memory.rs.
The large header comment from memorymap.rs describing the memory map
becomes a Markdown file.
